backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = "Backend for Kanban Board Fast API application."
authors = ["jakubpluta <plutakuba@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.78.0"
uvicorn = "^0.18.2"
tortoise-orm = {extras = ["asyncpg"], version = "^0.19.1"}
psycopg2-binary = "^2.9.3"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
python-multipart = "^0.0.5"
PyJWT = "^2.4.0"
python-dotenv = "^0.20.0"
python-jose = "^3.3.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
